ShapePath: Update `toShapes()`.
