+++ editor/js/Loader.js
@@ -716,7 +716,7 @@ function Loader( editor ) {
 
 						if ( fillMaterial ) {
 
-							const shapes = SVGLoader.createShapes( path );
+							const shapes = path.toShapes();
 
 							for ( let j = 0; j < shapes.length; j ++ ) {
