+++ examples/webgl_loader_svg.html
@@ -183,7 +183,7 @@
 
 								material.wireframe = guiData.fillShapesWireframe;
 
-								const shapes = SVGLoader.createShapes( path );
+								const shapes = path.toShapes();
 
 								for ( const shape of shapes ) {
